refactor(companhia-aerea): move todos os models para br.edu.ifto.sistemacompanhiaaerea.model
Refs: #72
