+++ sistema-companhia-aerea/src/main/java/br/edu/ifto/sistemacompanhiaaerea/model/MarcaAviao.java
@@ -1,4 +1,4 @@
-package br.edu.ifto.sistemacompanhiaaerea.entidades;
+package br.edu.ifto.sistemacompanhiaaerea.model;
 
 import javax.persistence.*;
 import javax.validation.constraints.NotBlank;
